Fix issue where client.loop floods the hardware
As noted in [this](https://github.com/knolleary/pubsubclient/issues/556#issuecomment-451673385) comment, there should
be a delay in the client.loop calls. As the sparsnas seems to only publish every 15 seconds
a 2.5 second delay should be sufficient whilst leaving plenty free cycles to do other stuff.

This will fix the issue in #30